[minor shit]

+++ Physics Engine/Physics Engine/Vector.h
@@ -93,4 +93,4 @@ template <typename dataType> class Vector
 	};
 
 typedef Vector <float> Vectorf;
-typedef Vector <double> Vectord;
+typedef Vector <float> Vectord;
